+++ src/main/java/ichinohe/application/bookmanagementsystem/service/core/EventRecordService.java
@@ -3,6 +3,7 @@ package ichinohe.application.bookmanagementsystem.service.core;
 import ichinohe.application.bookmanagementsystem.domain.alteration.BookAlterationApplication;
 import ichinohe.application.bookmanagementsystem.domain.core.BookEntity;
 import ichinohe.application.bookmanagementsystem.domain.core.BookRepository;
+import ichinohe.application.bookmanagementsystem.domain.core.EventRepository;
 import ichinohe.application.bookmanagementsystem.domain.delete.BookDeleteApplication;
 import ichinohe.application.bookmanagementsystem.domain.entry.BookEntryApplication;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,18 +16,20 @@ import org.springframework.stereotype.Service;
 public class EventRecordService {
     @Autowired
     private BookRepository bookRepository;
+    @Autowired
+    private EventRepository eventRepository;
 
     public void entry(BookEntryApplication application) {
         BookEntity bookEntity = bookRepository.findAliveBook(application);
-        bookRepository.entryEventRecord(bookEntity);
+        eventRepository.entryEventRecord(bookEntity);
     }
 
     public void delete(BookDeleteApplication application) {
-        bookRepository.deleteEventRecord(application);
+        eventRepository.deleteEventRecord(application);
     }
 
     public void alteration(BookAlterationApplication application) {
-        bookRepository.alterationEventRecord(application);
+        eventRepository.alterationEventRecord(application);
     }
 
 }
